+++ src/context/time.rs
@@ -28,14 +28,14 @@ impl ContextTime<'_> {
 /// Type of time resource to use.
 #[derive(Reflect, Debug, Default, Clone, Copy)]
 pub enum TimeKind {
+    /// Corresponds to [`Time<Real>`].
+    ///
+    /// Real wall-clock time elapsed, not affected by pausing or scaling.
+    #[default]
+    Real,
     /// Corresponds to [`Time`], which contains [`Time<Virtual>`], except in the fixed schedule,
     /// where it's [`Time<Fixed>`].
     ///
     /// Virtual game time, affected by [`Time::pause`] and [`Time::relative_speed`].
-    #[default]
     Virtual,
-    /// Corresponds to [`Time<Real>`].
-    ///
-    /// Real wall-clock time elapsed, not affected by pausing or scaling.
-    Real,
 }